refactor(mcp): shrink public API surface — underscore internal modules
Addresses Manoel's pre-merge review: the snapshot was committing posthog.mcp to a
large accidental public API (all ~22 implementation modules exposed, 343 lines).

- Underscore-prefix the 16 pure-internal modules (_sink, _instrumentation,
  _internal, _sanitization, _truncation, _capture, _ids, _intent, _exceptions,
  _compatibility, _posthog_events, _context_parameters, _conversation_id,
  _event_types, _instrument_fastmcp, _instrument_lowlevel) so they drop off the
  public surface entirely.
- Keep the modules that hold genuinely public symbols (constants, logger, session,
  tools, types, posthog_mcp, version) but add __all__ so each exposes only its
  intended public names. Underscore-prefixing these would have hidden the public
  classes from the snapshot too (it follows alias targets), so the guard would stop
  tracking the real API — __all__ keeps them tracked.

Public surface: 343 -> 95 lines, now exactly the __init__ __all__ set plus the
public types/enums. PostHogMCP's methods are still tracked. No runtime API change.

Generated-By: PostHog Code
Task-Id: b21bc954-5de3-4512-a0d5-6bec2371f782
