+++ posthog/mcp/_compatibility.py
