+++ posthog/mcp/_exceptions.py
