+++ posthog/mcp/_sanitization.py
