+++ posthog/mcp/_truncation.py
