+++ posthog/mcp/tools.py
@@ -12,6 +12,8 @@ from typing import Any, Dict, Optional
 
 from .logger import log
 
+__all__ = ["get_more_tools_result"]
+
 GET_MORE_TOOLS_NAME = "get_more_tools"
 
 _GET_MORE_TOOLS_RESULT_TEXT = (
